Moved CaseInsensitiveDict to utils
